requirements.txt
requests
pytz
sqlalchemy
pymysql
python-jose
fastapi[standard]
passlib
bcrypt
email-validator
python-ulid[pydantic]
ccxt
python-dotenv
redis[hiredis]